+++ changelog.md
@@ -91,7 +91,7 @@ fix:
     2.fix a bug that only can upload max 100 files
     3.fix a bug in generate
 
-v1.9.1_test
+v1.9.1
 add:
    1. add local IDE user type update fun
    2. add blog comment  fun
@@ -111,5 +111,12 @@ fix:
 2017.5.3 
 fix:
     1. blank command bug
+    2. page's tag is null when open a project
+    3. undo's bug when delete a page
 add:
     1. MyAnimation rendering
+    2. tag's register limit
+    3. anmiation's duration time limit
+    4. project's name limit when saveas
+
+